refactor: importing shared types to Problem, repositioning DefaultGoal and refactoring shared types.
