+++ public/ts/types/game.types.ts
@@ -1,13 +1,5 @@
-import { SlotCoords } from "./shared.types"
-
 export type PQueueItem<T> = {
     element: T
     cost: number
     parent?: T
-}
-
-export type GameResponse = {
-    message: string
-    success: boolean
-    solution?: Array<SlotCoords>
 }
