+++ public/ts/types/shared.types.ts
@@ -4,3 +4,9 @@ export type SlotCoords = {
     x: number,
     y: number
 }
+
+export type GameResponse = {
+    message: string
+    success: boolean
+    solution?: Array<SlotCoords>
+}
